requirements.txt
beautifulsoup4
pandas
requests
SQLAlchemy==1.4.47
wget
pymysql
python-dotenv